Refactor Smtp Configuration Resource
