+++ source/Octopus.Client.Tests/PublicSurfaceAreaFixture.ThePublicSurfaceAreaShouldNotRegress..NETCore.approved.txt
@@ -5451,27 +5451,6 @@ Octopus.Client.Model
       None = 0
       SkipUnavailableMachines = 1
   }
-  class SmtpConfigurationResource
-    Octopus.Client.Extensibility.IResource
-    Octopus.Client.Model.IAuditedResource
-    Octopus.Client.Model.Resource
-  {
-    .ctor()
-    String AzureAudience { get; set; }
-    String AzureClientId { get; set; }
-    String AzureTenantId { get; set; }
-    String CredentialType { get; set; }
-    Boolean EnableSsl { get; set; }
-    String GoogleAudience { get; set; }
-    String GoogleServiceAccount { get; set; }
-    String NewSmtpPassword { get; set; }
-    String SendEmailFrom { get; set; }
-    String SmtpHost { get; set; }
-    String SmtpLogin { get; set; }
-    Octopus.Client.Model.SensitiveValue SmtpPassword { get; set; }
-    Nullable<Int32> SmtpPort { get; set; }
-    Nullable<Int32> Timeout { get; set; }
-  }
   class SnapshotGitReferenceResource
   {
     .ctor()
@@ -7444,6 +7423,51 @@ Octopus.Client.Model.Migrations
     String SpaceId { get; set; }
     String SuccessCallbackUri { get; set; }
     String TaskId { get; set; }
+  }
+}
+Octopus.Client.Model.SmtpConfiguration
+{
+  class AzureSmtpConfigurationResource
+    Octopus.Client.Extensibility.IResource
+    Octopus.Client.Model.IAuditedResource
+    Octopus.Client.Model.SmtpConfigurationResource
+  {
+    .ctor()
+    String AzureAudience { get; set; }
+    String AzureClientId { get; set; }
+    String AzureTenantId { get; set; }
+  }
+  class GoogleSmtpConfigurationResource
+    Octopus.Client.Extensibility.IResource
+    Octopus.Client.Model.IAuditedResource
+    Octopus.Client.Model.SmtpConfigurationResource
+  {
+    .ctor()
+    String GoogleAudience { get; set; }
+    String GoogleServiceAccount { get; set; }
+  }
+  class SmtpConfigurationResource
+      Octopus.Client.Extensibility.IResource
+      Octopus.Client.Model.IAuditedResource
+      Octopus.Client.Model.Resource
+    {
+      .ctor()
+      String SmtpHost { get; set; }
+      Nullable<Int32> SmtpPort { get; set; }
+      String SendEmailFrom { get; set; }
+      Boolean EnableSsl { get; set; }
+      Nullable<Int32> Timeout { get; set; }
+      String CredentialType { get; set; }
+    }
+  class UsernamePasswordSmtpConfigurationResource
+    Octopus.Client.Extensibility.IResource
+    Octopus.Client.Model.IAuditedResource
+    Octopus.Client.Model.SmtpConfigurationResource
+  {
+    .ctor()
+    String SmtpLogin { get; set; }
+    Octopus.Client.Model.SensitiveValue SmtpPassword { get; set; }
+    String NewSmtpPassword { get; set; }
   }
 }
 Octopus.Client.Model.TenantVariables
